set up grabbing profiles on Home.tsx
